hotfix

+++ lightning_modules.py
@@ -557,7 +557,7 @@ class L_ResNext50(ptl.LightningModule):
         :param x:
         :return:
         """
-        x = x.unsqueeze(1)
+        x = x.unsqueeze(1).float()
         return self.model(x)
 
     def training_step(self, batch, batch_idx):
